Fix(front) : 회원 정보 조회, 수정 UI 변경 및 검증 로직 추가 / package경로 오류 수정  (S14P11D102-300)

+++ FrontEnd/app/src/main/java/com/ssafy/closetory/dto/SignUpRequest.kt
@@ -12,7 +12,7 @@ data class SignUpRequest(
     val passwordConfirm: String,
     val nickname: String,
     val gender: String, // "Male" or "Female"
-    val height: Int, // cm
-    val weight: Int, // kg
+    val height: Short, // cm
+    val weight: Short, // kg
     val alarmEnabled: Boolean
 )
